+++ src/commands/Plan.js
@@ -6,7 +6,7 @@ import _ from 'lodash'
 class Plan extends Command {
   main (cargo, cb) {
     if (!_.has(this.runtime.config, 'infra')) {
-      this.info(`Skipping as there are no install instructions\n`)
+      this._out(`Skipping as there are no install instructions\n`)
       return cb(null)
     }